Deleting now only possible by auth user

+++ part4/blogs/backend/controllers/blogs.js
@@ -2,6 +2,7 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
+const blog = require('../models/blog')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -35,6 +36,18 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  console.log(decodedToken)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const userid = decodedToken.id.toString()
+  const blog = await Blog.findById(request.params.id)
+
+  if (!(userid === blog.user.toString())){
+    return response.status(401).json({ error: 'post not created by this user' })
+  }
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
